Rename checkBlocks.js to prePrase.js then to prepareInput.js
Now prepareInput is a dependency passed to pegjs compiler, it is imported, it is standalone module that perform some mutations in the input latex to be parsed, such as remove unnecessary braces {}, these braces will not do effect after \\cos for example. There is too modification to text and peg$computePosDetails functions.